pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "air-relay"
version = "0.1.0"
description = "SMS to Telegram Bridge using Luat Air780E"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [
    { name = "AirRelay Team" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Communications :: Telephony",
]
dependencies = [
    "telethon>=1.27.0",
    "paho-mqtt>=2.0.0",
    "workers-kv.py>=1.2.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.18.0",
    "black>=23.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
    "autoflake>=2.0.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/air-relay"
"Bug Tracker" = "https://github.com/yourusername/air-relay/issues"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.18.0",
    "black>=23.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
    "autoflake>=2.0.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
asyncio_mode = "auto"
testpaths = ["tests"] 